Fix float32 truncation in balanced_allocation iteration loop (#1203)
_allocate_biased and _allocate_from_costs cast the allocation array to
float32 internally.  The balancing loop then compared these float32
values against float64 source IDs.  For non-integer IDs (e.g. 0.1, 0.3)
the float64->float32->float64 round-trip produces a different value, so
alloc == sid never matched, weights were all zero, and the loop exited
on the first iteration without balancing.

Fix: keep the allocation array as float64 through the iteration loop.
The final cast to float32 happens at the return statement.

+++ xrspatial/balanced_allocation.py
@@ -158,7 +158,7 @@ def _allocate_from_costs(cost_stack, source_ids, fill_value=np.nan):
     else:
         all_nan = np.all(np.isnan(np.stack(cost_stack, axis=0)), axis=0)
 
-    alloc = alloc.astype(np.float32)
+    alloc = alloc.astype(np.float64)
     alloc[all_nan] = fill_value
 
     return alloc
@@ -199,7 +199,7 @@ def _allocate_biased(cost_stack, biases, source_ids, fill_value=np.nan):
         stacked = np.stack(layers, axis=0)
         best_idx = np.argmin(stacked, axis=0)
 
-    alloc = source_ids[best_idx].astype(np.float32)
+    alloc = source_ids[best_idx].astype(np.float64)
 
     # Mark unreachable cells
     if da is not None and isinstance(first, da.Array):
